05_july_2025_14:00est home.spec.js, helpers.js are updated and products_en_ru.js is added
