+++ tests/ui/tests/home.spec.js
@@ -1,106 +1,8 @@
 // E:\Gurov_SSD_256\IT\Testing\Automation_08_09_2019\zakfarm\tests\ui\tests\home.spec.js
 const { test, expect } = require('@playwright/test');
 const HomePage = require('../pages/homePage');
-
-// Combined product list with language-specific names
-const PRODUCTS = [
-  { en_name: 'Box Premium', ru_name: 'Box Premium', price: '$348.00' },
-  { en_name: 'Box Gold', ru_name: 'Бокс Gold', price: '$256.00' },
-  { en_name: 'Box Standart', ru_name: 'Бокс Standart', price: '$158.00' },
-  { en_name: 'Hallumi With Provencal Herbs', ru_name: 'Халуми С Прованскими Травами', price: '$45.00' },
-  { en_name: 'Hallumi', ru_name: 'Халлуми', price: '$45.00' },
-  { en_name: 'Cheese Sticks (Cecil Cheese)', ru_name: 'Сырные Палочки (Сыр Чечил)', price: '$40.00' },
-  { en_name: 'Suluguni Cheese (Head)', ru_name: 'Сулугун Сыр', price: '$40.00' },
-  { en_name: 'Ossetian Cheese (Brynza). Middle-Salty, Ripe.', ru_name: 'Осетинский Сыр (Брынза) Среднесоленый, Выдержанный', price: '$45.00' },
-  { en_name: 'NEW Cheese Plate Ossetian Cheese (Brynza) 170-200gr, Adygean Cheese 150 Gr, Cheese Sticks (Cecil Cheese) 120', ru_name: 'Сырная Тарелка', price: '$40.00' },
-  { en_name: 'Adygean Cheese', ru_name: 'Адыгейский Сыр', price: '$45.00' },
-  { en_name: 'Ossetian Cheese (Brynza) Slightly Salty.', ru_name: 'Осетинский Сыр (Брынза) Слабосоленый', price: '$45.00' },
-  { en_name: 'Pancakes With Meat', ru_name: 'Блинчики С Мясом', price: '$20.00' },
-  { en_name: 'Pancakes With Cottage Cheese', ru_name: 'Блинчики С Творогом', price: '$20.00' },
-  { en_name: 'Сottage Cheesecake Gluten Free', ru_name: 'Сырники Gluten Free', price: '$25.00' },
-  { en_name: 'Cottage Cheesecake With Chocolate', ru_name: 'Сырники С Шоколадом', price: '$20.00' },
-  { en_name: 'Сottage Cheesecake With Poppy', ru_name: 'Сырники С Маком', price: '$20.00' },
-  { en_name: 'Сottage Cheesecake With Raisins', ru_name: 'Сырники С Изюмом', price: '$20.00' },
-  { en_name: 'Georgian Cottage Cheesecake', ru_name: 'Сырники «По-Грузински»', price: '$23.00' },
-  { en_name: 'Сottage', ru_name: 'Сырники', price: '$20.00' },
-  { en_name: 'Matsoni', ru_name: 'Мацони', price: '$15.00' },
-  { en_name: 'Melted Butter', ru_name: 'Топленое Масло (Гхи)', price: '$50.00' },
-  { en_name: 'Homemade Butter', ru_name: 'Домашнее Сливочное Масло', price: '$30.00' },
-  { en_name: 'Сlassic Sour Cream', ru_name: 'Классическая Сметана', price: '$18.00' },
-  { en_name: 'Cottage Cheese With Fat Up To 5%', ru_name: 'Творог Жирностью До 5%', price: '$18.00' },
-  { en_name: 'Lobiani', ru_name: 'Лобиани', price: '$30.00' },
-  { en_name: 'Ossetian Cherry Pie', ru_name: 'Осетинский Пирог С Вишней', price: '$40.00' },
-  { en_name: 'Ossetian Pie With Potatoes And Cheese', ru_name: 'Осетинский Пирог С Картофелем И Сыром', price: '$40.00' },
-  { en_name: 'Ossetian Meat Pie (Beef)', ru_name: 'Осетинский Пирог С Мясом (Говядина)', price: '$50.00' },
-  { en_name: 'Ossetian Pie With Cheese And Beet Leaves.', ru_name: 'Осетинский Пирог С Сыром И Листьями Свеклы', price: '$40.00' },
-  { en_name: 'Ossetian Pie With Cheese', ru_name: 'Осетинский Пирог С Сыром', price: '$40.00' },
-  { en_name: 'Cheese Khinkali', ru_name: 'Хинкали С Сыром', price: '$50.00' },
-  { en_name: 'Cheese And Mint Khinkali', ru_name: 'Хинкали С Сыром И Мятой', price: '$50.00' },
-  { en_name: 'Beef & Lamb Khinkali', ru_name: 'Хинкали С Говядиной И Бараниной', price: '$50.00' },
-  { en_name: 'Dumplings', ru_name: 'Пельмени', price: '$45.00' },
-  { en_name: 'Khinkali (Meat Dumplings)', ru_name: 'Хинкали', price: '$45.00' },
-  { en_name: 'Vareniki With Adygean Cheese', ru_name: 'Вареники С Адыгейским Сыром', price: '$20.00' },
-  { en_name: 'Vareniki With Cottage Cheese', ru_name: 'Вареники С Творогом', price: '$20.00' },
-  { en_name: 'Vareniki With Potatoes', ru_name: 'Вареники С Картофелем', price: '$15.00' },
-  { en_name: 'Vareniki With Cherries', ru_name: 'Вареники С Вишней', price: '$20.00' },
-  { en_name: 'Khachapuri In Adjarian', ru_name: 'Хачапури По-Аджарски', price: '$25.00' },
-  { en_name: 'Pork Beer Sausages', ru_name: 'Колбаски Из Свинины Под Пиво', price: '$12.00' },
-  { en_name: 'Krakowska Sausage', ru_name: 'Краковская Колбаса', price: '$26.00' },
-  { en_name: 'Smoked Pork (Carbonade)', ru_name: 'Копчёная Свинина (Карбонат)', price: '$22.00' },
-  { en_name: 'Smoked Chicken Drumsticks (With Skin)', ru_name: 'Копчёные Куриные Голени С Кожей', price: '$16.00' },
-  { en_name: 'Beef Loin TOP Sirloin Steak', ru_name: 'Стейк Говяжий Топ-Сирлойн', price: '$85.00' },
-  { en_name: 'Beef Steak', ru_name: 'Говяжий Стейк', price: '$125.00' },
-  { en_name: 'Breaded Chicken Cutlets 1 Lb', ru_name: 'Котлеты Куриные В Панировке 1 Lb', price: '$20.00' },
-  { en_name: 'Chicken Cutlets 1 Lb', ru_name: 'Котлеты Куриные 1 Lb', price: '$20.00' },
-  { en_name: 'Breaded Pork/Beef Cutlets 1 Lb', ru_name: 'Котлеты Свинина/Говядина В Панировке 1 Lb', price: '$25.00' },
-  { en_name: 'Pork/Beef Cutlets 1 Lb', ru_name: 'Котлеты Свинина/Говядина 1 Lb', price: '$25.00' },
-  { en_name: 'Free-Range Chicken Egg', ru_name: 'Яйцо Домашнее', price: '$10.00' },
-  { en_name: 'Organic Poultry Chicken (Frozen)', ru_name: 'Куриное Мясо Органическое (Замороженное)', price: '$40.00' },
-  { en_name: 'Ground Turkey Chub', ru_name: 'Фарш Индюшачий', price: '$32.00' },
-  { en_name: 'Extra Lean Ground Turkey', ru_name: 'Фарш Индюшачий Постный', price: '$34.00' },
-  { en_name: 'Ground Pork (Frozen)', ru_name: 'Фарш Свиной (Замороженный)', price: '$25.00' },
-  { en_name: 'Pork Ham (Frozen)', ru_name: 'Свиной Окорок', price: '$20.00' },
-  { en_name: 'Rack Of Lamb (Grass-Fed)', ru_name: 'Каре Ягненка (Откормленного На Траве)', price: '$75.00' },
-  { en_name: 'Leg Of Lamb (Grass-Fed) (Frozen)', ru_name: 'Нога Ягненка (Откормленного На Траве)', price: '$75.00' },
-  { en_name: 'ABKHAZURA Georgian Meatballs (Frozen)', ru_name: 'АБХАЗУРА Грузинские Мясные Котлеты (Замороженные)', price: '$40.00' },
-  { en_name: 'Organic Goose, Whole (Frozen)', ru_name: 'Органический Гусь Цельный (Замороженный)', price: '$115.00' },
-  { en_name: 'Farm Duck (Frozen)', ru_name: 'Утка Фермерская (В Замороженном Виде)', price: '$90.00' },
-  { en_name: 'Pasture-Raised Turkey (Frozen)', ru_name: 'Индюшка, Выращенная На Пастбище (Замороженная)', price: '$260.00' },
-  { en_name: 'Beef Jerky', ru_name: 'Джерки Вяленая Говядина', price: '$20.00' },
-  { en_name: 'Smoked Beef Ribs', ru_name: 'Ребра Говяжьи Горячего Копчения', price: '$90.00' },
-  { en_name: 'Hot Smoked Pork Briskets', ru_name: 'Грудинка Свиная Горячего Копчения', price: '$40.00' },
-  { en_name: 'Spinach Pkhali', ru_name: 'Пхали Из Шпината', price: '$40.00' },
-  { en_name: 'Beet Leaf Pkhali 1 Lb', ru_name: 'Пхали Из Свекольных Листьев', price: '$40.00' },
-  { en_name: 'Ajapsandali (Frozen)', ru_name: 'Аджапсандали (Замороженный)', price: '$20.00' },
-  { en_name: 'Chakhokhbili (Frozen)', ru_name: 'Чахохбили (Замороженное)', price: '$20.00' },
-  { en_name: 'Eggplant Rolls (Frozen)', ru_name: 'Рулетики Из Баклажанов (Замороженные)', price: '$25.00' },
-  { en_name: 'Lobio (Frozen)', ru_name: 'Лобио (Замороженное)', price: '$20.00' },
-  { en_name: 'Churchkhela – Walnut, Prune & Dried Apricot', ru_name: 'Чурчхела – Грецкий Орех, Чернослив И Курага', price: '$30.00' },
-  { en_name: 'Churchkhela Assorted (3 Pcs)', ru_name: 'Чурчхела Ассорти – Грецкий Орех, Фундук, Грецкий', price: '$30.00' },
-  { en_name: 'Churchkhela – Fig & Prune (3 Pcs)', ru_name: 'Чурчхела – Инжир И Чернослив', price: '$30.00' },
-  { en_name: 'Churchkhela – Walnut (3 Pcs)', ru_name: 'Чурчхела – Грецкий Орех', price: '$30.00' },
-  { en_name: 'Churchkhela – Prune & Walnut (3 Pcs)', ru_name: 'Чурчхела – Чернослив И Грецкий Орех', price: '$30.00' },
-  { en_name: 'Raspberry Pastilles', ru_name: 'Малиновая Пастила.', price: '$6.00' },
-  { en_name: 'Pineapple, Mango Pastilles', ru_name: 'Ананас-Манго Пастила', price: '$6.00' },
-  { en_name: 'Cherry Pastilles', ru_name: 'Вишневая Пастила.', price: '$5.00' },
-  { en_name: 'Seabuckthorn Pastilles', ru_name: 'Облепиховая Пастила.', price: '$6.00' },
-  { en_name: 'Red Currant Pastilles', ru_name: 'Пастила Из Красной Смородины.', price: '$6.00' },
-  { en_name: 'Black Currant Pastilles', ru_name: 'Пастила Из Черной Смородины.', price: '$6.00' },
-  { en_name: 'Strawberry Mint Pastilles', ru_name: 'Клубничная Пастила С Мятой.', price: '$6.00' },
-  { en_name: 'Chocolate Pastilles', ru_name: 'Шоколадная Пастила.', price: '$6.00' },
-  { en_name: 'Banana Pastilles', ru_name: 'Банановая Пастила.', price: '$6.00' },
-  { en_name: 'Strawberry Pastilles', ru_name: 'Клубничная Пастила.', price: '$6.00' },
-  { en_name: 'Apple Pastilles', ru_name: 'Яблочная Пастила.', price: '$6.00' },
-  { en_name: 'Black Forest Cake', ru_name: 'Торт Черный Лес (Black Forest)', price: '$80.00' },
-  { en_name: 'Mango-Passion Fruit Cake', ru_name: 'Торт Манго-Маракуйя', price: '$80.00' },
-  { en_name: 'Kiev Cake', ru_name: 'Киевский Торт', price: '$65.00' },
-  { en_name: 'Napoleon Cake', ru_name: 'Наполеон', price: '$65.00' },
-  { en_name: 'Honey Cake', ru_name: 'Медовик', price: '$60.00' }
-];
-
-function normalize(str) {
-  return str.trim().toLowerCase().replace(/\s+/g, ' ');
-}
+const { PRODUCTS } = require('../utils/products_en_ru');
+const Helpers = require('../utils/helpers');
 
 async function runProductDetailsTest(page, baseUrl, language) {
     const API_TIMEOUT = process.env.API_TIMEOUT ? parseInt(process.env.API_TIMEOUT) : 30000;
@@ -124,7 +26,7 @@ async function runProductDetailsTest(page, baseUrl, language) {
         
         try {
             const match = allProducts.find(actual =>
-                normalize(actual.name) === normalize(expected[nameKey]) &&
+                Helpers.normalize(actual.name) === Helpers.normalize(expected[nameKey]) &&
                 actual.price.trim() === expected.price.trim()
             );
